+++ pyrates/frontend/file.py
@@ -31,7 +31,7 @@
 import importlib
 
 from pyrates import PyRatesException
-from pyrates.frontend import yaml as _yaml
+from pyrates.frontend.fileio import yaml as _yaml
 
 __author__ = "Daniel Rose"
 __status__ = "Development"
@@ -53,7 +53,8 @@ file_loader_mapping = {"yaml": _yaml.to_dict,
 
 
 def parse_path(path: str):
-    """Parse a path of form path.to.template, returning a tuple of (name, file, abspath)."""
+    """Parse a path of form path.to.template_file.template_name or path/to/template_file/template_name,
+    returning a tuple of (name, file, abspath)."""
 
     if "/" in path or "\\" in path:
         import os