+++ pyrates/frontend/template/circuit/circuit.py
@@ -39,6 +39,7 @@ from copy import deepcopy
 
 # pyrates internal imports
 from pyrates import PyRatesException
+from pyrates.frontend.template._io import _complete_template_path
 from pyrates.frontend.template.abc import AbstractBaseTemplate
 from pyrates.frontend.template.edge import EdgeTemplate
 from pyrates.frontend.template.node import NodeTemplate
@@ -65,7 +66,7 @@ class CircuitTemplate(AbstractBaseTemplate):
         if nodes:
             for key, node in nodes.items():
                 if isinstance(node, str):
-                    path = self._complete_template_path(node, self.path)
+                    path = _complete_template_path(node, self.path)
                     self.nodes[key] = NodeTemplate.from_yaml(path)
                 else:
                     self.nodes[key] = node
@@ -74,7 +75,7 @@ class CircuitTemplate(AbstractBaseTemplate):
         if circuits:
             for key, circuit in circuits.items():
                 if isinstance(circuit, str):
-                    path = self._complete_template_path(circuit, self.path)
+                    path = _complete_template_path(circuit, self.path)
                     self.circuits[key] = CircuitTemplate.from_yaml(path)
                 else:
                     self.circuits[key] = circuit
@@ -235,23 +236,27 @@ class CircuitTemplate(AbstractBaseTemplate):
 
             if isinstance(edge, dict):
                 try:
-                    edge = (edge["source"], edge["target"], edge["template"], edge["variables"])
+                    source = edge["source"]
+                    target = edge["target"]
+                    template = edge["template"]
+                    variables = edge["variables"]
                 except KeyError as e:
                     raise TypeError(f"Wrong edge configuration. Unable to find key {e.args[0]}")
 
+            else:
+                source, target, template, variables = edge
+
             # "template" is EdgeTemplate, just use it
-            if isinstance(edge[2], EdgeTemplate):
-                temp = edge[2]
+            # also just leave it as None, if so it be.
+            if isinstance(template, EdgeTemplate) or template is None:
+                pass
 
             # if not, try to load template path from yaml
             else:
-                path = edge[2]
-                try:
-                    path = self._complete_template_path(path, self.path)
-                    temp = EdgeTemplate.from_yaml(path)
-                except TypeError:
-                    temp = None
-            edges_with_templates.append((*edge[0:2], temp, edge[3]))
+                path = _complete_template_path(template, self.path)
+                template = EdgeTemplate.from_yaml(path)
+
+            edges_with_templates.append((source, target, template, variables))
         return edges_with_templates